QOL changes

+++ cogs/general.py
@@ -29,13 +29,17 @@ class GeneralCommands(commands.Cog):
         await ctx.channel.purge(limit=amount)
 
     @commands.command(name='bug',
-              aliases=['report', 'bugreport', 'br'],
-              decription="Send a bug report")
+                      aliases=['report', 'bugreport', 'br'],
+                      decription="Send a bug report")
     async def bug(self, ctx, *args):
-        feedbackChannel = self.bot.get_channel(779738194097995806)
-        await feedbackChannel.send(f"**BUG REPORT**\n{ctx.author.id} (**{ctx.message.author.name}#{ctx.message.author.discriminator}**) says ```css\n{' '.join(args)}```")
-        await ctx.message.delete()
-        await ctx.send("Thank you for that")
+        if (len(args) == 0):
+            await ctx.send("Type `!bug <bug you encountered>`")
+            return
+        else:
+            feedbackChannel = self.bot.get_channel(779738194097995806)
+            await feedbackChannel.send(f"**BUG REPORT**\n{ctx.author.id} (**{ctx.message.author.name}#{ctx.message.author.discriminator}**) says ```css\n{' '.join(args)}```")
+            await ctx.message.delete()
+            await ctx.send("Thank you for that")
 
 
 def setup(bot):
